Move array.js to core/types
