+++ extensions/amp-a4a/0.1/test/test-amp-a4a-within-viewport.js
@@ -22,7 +22,7 @@ import '../../../../extensions/amp-ad/0.1/amp-ad-xorigin-iframe-handler';
 import '../../../amp-ad/0.1/amp-ad';
 import {AmpA4A} from '../amp-a4a';
 import {htmlFor} from '../../../../src/static-template';
-import {pushIfNotExist, removeItem} from '../../../../src/utils/array';
+import {pushIfNotExist, removeItem} from '../../../../src/core/types/array';
 import {upgradeOrRegisterElement} from '../../../../src/service/custom-element-registry';
 
 describes.realWin(
