+++ extensions/amp-bind/0.1/bind-evaluator.js
@@ -17,7 +17,7 @@
 import {BindExpression} from './bind-expression';
 import {BindMacro} from './bind-macro';
 import {BindValidator} from './bind-validator';
-import {remove} from '../../../src/utils/array';
+import {remove} from '../../../src/core/types/array';
 
 /**
  * Asynchronously evaluates a set of Bind expressions.
