+++ extensions/amp-story-auto-ads/0.1/story-ad-page-manager.js
@@ -27,7 +27,7 @@ import {
 import {StoryAdLocalization} from './story-ad-localization';
 import {StoryAdPage} from './story-ad-page';
 import {devAssert} from '../../../src/log';
-import {findIndex} from '../../../src/utils/array';
+import {findIndex} from '../../../src/core/types/array';
 import {getServicePromiseForDoc} from '../../../src/service';
 
 /** @const {string} */
