+++ src/context/node.js
@@ -17,7 +17,7 @@
 import {Values} from './values';
 import {devAssert} from '../log';
 import {getMode} from '../mode';
-import {pushIfNotExist, removeItem} from '../utils/array';
+import {pushIfNotExist, removeItem} from '../core/types/array';
 import {throttleTail} from './scheduler';
 
 // Properties set on the DOM nodes to track the context state.
