+++ src/core/types/array.js
