+++ src/gesture.js
@@ -17,7 +17,7 @@
 import {Observable} from './observable';
 import {Pass} from './pass';
 import {devAssert} from './log';
-import {findIndex} from './utils/array';
+import {findIndex} from './core/types/array';
 import {supportsPassiveEventListener} from './event-helper-listen';
 import {toWin} from './types';
