+++ test/unit/test-amp-inabox.js
@@ -18,7 +18,7 @@ import * as Service from '../../src/service';
 import {AmpDocSingle} from '../../src/service/ampdoc-impl';
 import {installAmpdocServices} from '../../src/service/core-services';
 import {installAmpdocServicesForInabox} from '../../src/inabox/inabox-services';
-import {removeItem} from '../../src/utils/array';
+import {removeItem} from '../../src/core/types/array';
 
 describe('amp-inabox', () => {
   describes.realWin('installAmpdocServicesForInabox', {amp: false}, (env) => {
