+++ test/unit/utils/test-display-observer.js
@@ -22,7 +22,7 @@ import {
   unobserveDisplay,
   unregisterContainer,
 } from '../../../src/utils/display-observer';
-import {removeItem} from '../../../src/utils/array';
+import {removeItem} from '../../../src/core/types/array';
 
 describes.realWin('display-observer', {amp: true}, (env) => {
   let win, doc, ampdoc;
